implemented shared preferences

+++ app/src/main/java/com/tapatuniforms/pos/network/ProductAPI.java
@@ -18,6 +18,7 @@ import com.tapatuniforms.pos.helper.APIErrorListener;
 import com.tapatuniforms.pos.helper.APIStatic;
 import com.tapatuniforms.pos.helper.DatabaseSingleton;
 import com.tapatuniforms.pos.helper.NotifyListener;
+import com.tapatuniforms.pos.helper.SharedPrefs;
 import com.tapatuniforms.pos.helper.Validator;
 import com.tapatuniforms.pos.helper.VolleySingleton;
 import com.tapatuniforms.pos.model.Category;
@@ -53,6 +54,7 @@ public class ProductAPI {
     public static synchronized ProductAPI getInstance(Context context) {
         if (instance == null) {
             instance = new ProductAPI(context);
+            SharedPrefs.init(context);
         }
 
         return instance;
@@ -122,8 +124,6 @@ public class ProductAPI {
                                      InventoryOrderAdapter inventoryOrderAdapter, InventoryFragment inventoryFragment) {
 
         List<ProductHeader> localProductList = db.productHeaderDao().getAllProductHeader();
-//        if (!Validator.isNetworkConnected(context)) {
-//            Toast.makeText(context, context.getString(R.string.no_network), Toast.LENGTH_SHORT).show();
         productList.clear();
         productList.addAll(localProductList);
 
@@ -149,22 +149,18 @@ public class ProductAPI {
             inventoryOrderAdapter.notifyDataSetChanged();
         }
 
-//            return;
-//        }
-
         DjangoJSONArrayResponseRequest request = new DjangoJSONArrayResponseRequest(
                 Request.Method.GET, APIStatic.Outlet.productUrl, null,
                 response -> {
                     // Response Received
-                    if (response.length() != localProductList.size()) {
+                    //TODO: apply sync adapter
+                    if (response.length() != 0 && response.length() != localProductList.size()) {
                         for (int i = 0; i < response.length(); i++) {
                             JSONObject productJSON = response.optJSONObject(i);
                             ProductHeader product = new ProductHeader(productJSON);
 
                             productList.add(product);
                             allProducts.add(product);
-
-                            //TODO: check sync status, delete the item and insert it again
                         }
                         db.productHeaderDao().deleteAll();
                         db.productVariantDao().deleteAll();
@@ -179,28 +175,9 @@ public class ProductAPI {
                         }
                     }
 
-                    /*List<Stock> stockList = db.stockDao().getAll();
-                    if (stockList == null || stockList.size() < 1)
-                        for (ProductHeader product : allProducts) {
-                            List<ProductVariant> variantList = db.productVariantDao().getProductVariantsById(product.getId());
-
-                            for (ProductVariant variant : variantList) {
-                                Stock stock = new Stock(variant.getId(), variant.getDisplayStock(), variant.getWarehouseStock());
-                                db.stockDao().insert(stock);
-                            }
-                        }*/
-
                     if (inventoryAdapter != null) {
                         inventoryAdapter.notifyDataSetChanged();
                     }
-
-                    /*if (inventoryFragment != null) {
-                        inventoryFragment.getRecommendedProductList();
-                    }
-
-                    if (inventoryOrderAdapter != null) {
-                        inventoryOrderAdapter.notifyDataSetChanged();
-                    }*/
 
                     if (productAdapter != null) {
                         productAdapter.notifyDataSetChanged();
@@ -269,10 +246,7 @@ public class ProductAPI {
                     return;
                 }
 
-                List<Outlet> outletList = db.outletDao().getAll();
-                long outletId = -1;
-                if (outletList != null && outletList.size() > 0)
-                    outletId = outletList.get(0).getId();
+                long outletId = SharedPrefs.readInt(APIStatic.Outlet.OUTLET_ID, -1);
                 stock.setDisplay(productVariant.getDisplayStock() - subOrder.getQuantity());
 
                 JSONObject stockJson = stock.toJson();
@@ -434,10 +408,14 @@ public class ProductAPI {
 
     public void getOutletList(ArrayList<Outlet> outletList, DatabaseSingleton db, NotifyListener listener) {
 
-        List<Outlet> outletList1 = db.outletDao().getAll();
-        if (outletList1.size() > 0) {
-            outletList.clear();
-            outletList.add(outletList1.get(0));
+        if (!Validator.isNetworkConnected(context)) {
+            List<Outlet> outletList1 = db.outletDao().getAll();
+            if (outletList1.size() > 0) {
+                outletList.clear();
+                outletList.add(outletList1.get(0));
+
+                SharedPrefs.writeInt(APIStatic.Outlet.OUTLET_ID, outletList1.get(0).getId());
+            }
             return;
         }
 
@@ -451,7 +429,7 @@ public class ProductAPI {
                         JSONObject outletJSON = response.optJSONObject(0);
 
                         Outlet outlet = new Outlet(outletJSON);
-
+                        SharedPrefs.writeInt(APIStatic.Outlet.OUTLET_ID, outlet.getId());
                         ArrayList<School> schoolList = new ArrayList<>(db.schoolDao().getAll());
                         if (schoolList.size() < 1) {
                             SchoolAPI.getInstance(context).getSchool(schoolList, db, new NotifyListener() {
